refactor(server): improve FPS Stats collection logic
This commit extracts the FPS statistics collection into its own class to keep
the `VideoStreamTrack` implementation cleaner and more maintainable. This also
makes the logic reusable across different components.
