+++ server/utils.py
@@ -1,4 +1,5 @@
 """Utility functions for the server."""
+
 import asyncio
 import random
 import types
@@ -7,6 +8,8 @@ import json
 from aiohttp import web
 from aiortc import MediaStreamTrack
 from typing import List, Tuple, Any, Dict
+import time
+from collections import deque
 
 logger = logging.getLogger(__name__)
 
@@ -78,7 +81,9 @@ class StreamStats:
         """
         self._app = app
 
-    async def collect_video_metrics(self, video_track: MediaStreamTrack) -> Dict[str, Any]:
+    async def collect_video_metrics(
+        self, video_track: MediaStreamTrack
+    ) -> Dict[str, Any]:
         """Collects real-time statistics for a video track.
 
         Args:
@@ -88,10 +93,10 @@ class StreamStats:
             A dictionary containing FPS-related statistics.
         """
         return {
-            "timestamp": await video_track.last_fps_calculation_time,
-            "fps": await video_track.fps,
-            "minute_avg_fps": await video_track.average_fps,
-            "minute_fps_array": await video_track.fps_measurements,
+            "timestamp": await video_track.fps_meter.last_fps_calculation_time,
+            "fps": await video_track.fps_meter.fps,
+            "minute_avg_fps": await video_track.fps_meter.average_fps,
+            "minute_fps_array": await video_track.fps_meter.fps_measurements,
         }
 
     async def collect_all_stream_metrics(self, _) -> web.Response:
@@ -132,3 +137,97 @@ class StreamStats:
             content_type="application/json",
             text=json.dumps(stats),
         )
+
+
+class FPSMeter:
+    """Class to calculate and store the framerate of a stream by counting frames."""
+
+    def __init__(self):
+        """Initializes the FPSMeter class."""
+        self._lock = asyncio.Lock()
+        self._fps_interval_frame_count = 0
+        self._last_fps_calculation_time = None
+        self._fps_loop_start_time = None
+        self._fps = 0.0
+        self._fps_measurements = deque(maxlen=60)
+        self._running_event = asyncio.Event()
+
+        asyncio.create_task(self._calculate_fps_loop())
+
+    async def _calculate_fps_loop(self):
+        """Loop to calculate FPS periodically."""
+        await self._running_event.wait()
+        self._fps_loop_start_time = time.monotonic()
+        while True:
+            async with self._lock:
+                current_time = time.monotonic()
+                if self._last_fps_calculation_time is not None:
+                    time_diff = current_time - self._last_fps_calculation_time
+                    self._fps = self._fps_interval_frame_count / time_diff
+                    self._fps_measurements.append(
+                        {
+                            "timestamp": current_time - self._fps_loop_start_time,
+                            "fps": self._fps,
+                        }
+                    )  # Store the FPS measurement with timestamp
+
+                # Reset start_time and frame_count for the next interval.
+                self._last_fps_calculation_time = current_time
+                self._fps_interval_frame_count = 0
+            await asyncio.sleep(1)  # Calculate FPS every second.
+
+    async def increment_frame_count(self):
+        """Increment the frame count to calculate FPS."""
+        async with self._lock:
+            self._fps_interval_frame_count += 1
+            if not self._running_event.is_set():
+                self._running_event.set()
+
+    @property
+    async def fps(self) -> float:
+        """Get the current output frames per second (FPS).
+
+        Returns:
+            The current output FPS.
+        """
+        async with self._lock:
+            return self._fps
+
+    @property
+    async def fps_measurements(self) -> list:
+        """Get the array of FPS measurements for the last minute.
+
+        Returns:
+            The array of FPS measurements for the last minute.
+        """
+        async with self._lock:
+            return list(self._fps_measurements)
+
+    @property
+    async def average_fps(self) -> float:
+        """Calculate the average FPS from the measurements taken in the last minute.
+
+        Returns:
+            The average FPS over the last minute.
+        """
+        async with self._lock:
+            if not self._fps_measurements:
+                return 0.0
+            return sum(
+                measurement["fps"] for measurement in self._fps_measurements
+            ) / len(self._fps_measurements)
+
+    @property
+    async def last_fps_calculation_time(self) -> float:
+        """Get the elapsed time since the last FPS calculation.
+
+        Returns:
+            The elapsed time in seconds since the last FPS calculation.
+        """
+        async with self._lock:
+            if (
+                self._last_fps_calculation_time is None
+                or self._fps_loop_start_time is None
+            ):
+                return 0.0
+            return self._last_fps_calculation_time - self._fps_loop_start_time
